requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
ultralytics==8.2.0
Pillow==10.3.0
python-multipart==0.0.9
gdown==5.2.0
